2020-07-26
Sunday, July 26th, 2020 - the not a theme, a seam, and a tribute to Peter Green show! From Ani DiFranco to Peter Green, with some legendary musicians in between. I followed a seam or a thread to get from Ani to Peter by way of Bob Dylan, Leonard Cohen, Phil Spector, The Beatles, Billy Preston, The Rolling Stones, the holy triumvirate of Eric Clapton, Jeff Beck and Jimmy Page, closing out with a selection of Peter Green's music with John Mayall, the Splinter Group and Fleetwood Mac.