2020-07-28
Tuesday, July 28th, 2020 - the female singer-songwriter show! I wouldn't stretch to calling this a Lilith Fair show (mainly for copyright reasons), but it is certainly a showcase of some of my favourite female singer-songwriters from the late 80s to present day - most of whom have played Lilith Fair! If that is a celebration of women in music, then I'm all for it!