2020-07-29
Wednesday, July 29th, 2020 - the muscle-car rock show! Inspired by an article I read on Tidal, this is the sound of early 1970s blues-based stadium hard rock. If your roof is down, and the road is open, this might be the show for you. Fire up that V8 and start munching miles!