2020-07-30
Thursday, July 30th, 2020 - the musical theatre show! Something of a trip down memory lane with selections from my favourite musicals that I worked on or performed in at Playhouse on the Square in Memphis, TN. So many great songs to pick from, and so many more to still play out. Consider this Act 1 of this theme, with more to come!