2020-07-31
Friday, July 31st, 2020 - the Phil's musical coming of age show - part 4! Opening with some catch-up tracks from the 1970s, before rejoining the musical autobiography of Phil "Dr Philphy" Edwards in 1989. This is the fourth part of the story of growing up to music in England and picks up right after the Atom Spies split up.