2020-08-13
Thursday, August 13th, 2020 - the "other" 1980s show! A little alternative, a little indie, some pop and some rock. Diverse, extravagant at times, and eclectic. This is a selection of some of the "other" music from the 1980s. Not hair metal, not power ballads, and not (as wonderful as it is) the bubblegum pop of the decade. What a weird and wonderful decade it was.