2020-08-14
Friday, August 14th, 2020 - the day in the life of 1994 show! 25 years ago today Blur and Oasis released singles on the same day - Country House and Roll With It. It was considered the great brit pop war, but what else was happening in 1994 if you are looking for a snapshot of popular music across a number of genres? There's only one way to find out!