2020-08-30
Sunday, August 30th, 2020 - the swinging 6/8 summer Sunday show! With a caution to not approach this too academically, this is a show, where the music isn't in 4/4 time. 6/8 has a lovely swing to it, and just makes you want to move. Today is a play-through of songs you feel in 3! If I hadn't mentioned it, you might not have noticed, so just ignore me and swing away!