2020-09-01
Tuesday, September 1st, 2020 - the selections from 68 & 69 show! I just started watching the documentary "Woodstock: 3 Days of Peace & Music." Spoiler alert: It's excellent! In a similar way to how I felt affected by the documentary Laurel Canyon, I found it humanised an event that has been elevated to such a level that made it feel inaccessible in some ways. For today's show I wanted to listen to and explore what was happening in rock music around the time of Woodstock, and consider what music may have been playing on the tape decks of the cars of the people that travelled there.