2020-09-03
Thursday, September 3rd, 2020 - the 1979 show! I've been waiting for an opportunity to play Prince again, and in particular the 1979 version of "I Could Never Take the Place Of Your Man." Instead of just compiling a playlist from that point, I decided to have a listen to what was popular across a variety of genres from that year, and go on a little musical excursion into the past.