2020-09-13
Sunday, September 13th, 2020 - the Tamarin time machine birthday show! I found an old set list from a band I was in around 2000-2002 called Tamarin. It was a Boxing Day (the day after Christmas) show, and at that time we were playing 3 or 4 times a week as a tiny power-duo. There's some good stuff on the setlist, and great memories hearing it back again. (These are the originals, not the covers we did!)