2020-09-20
Sunday, September 20th, 2020 - the family affair show! Father and son Adrian and Dave Tongue collaborated to curate today's playlist, and it's a journey through rock and pop from the early 1960s to the 1990s. Poppy, rocky, psychedelic, upbeat and happy, with just a few moody and sinister tracks in there too.