2020-09-26
Saturday, September 26th, 2020 - the funking funky as funk show! Quite fortuitously, I happened upon a highly amusing YouTuber called Davie504 this week, and that was all it took to inspire a super funky, bass heavy show. That and watching Mallrats last night! Strap in, turn it up, and get ready for maximum SLAPP! This was also the 150th show!