2020-10-07
Wednesday, October 7th, 2020 - the Eddie Van Halen show! I found myself as shocked and saddened as many guitarists, and music lovers seemed to be at the news that Eddie Van Halen died yesterday. Today's show was an insight into his influences and an example of his style, based on an article I read about him in Rolling Stone from a few years ago. Starting right at the beginning with Les Paul, speaking to his admiration for Eric Clapton, and gathering together all of the ingredients that came together to make him the musician he was.