2020-10-22
Thursday, October 22nd, 2020 - the Spencer Davis inspired show! Sadly, the inspiration for this show began with another obituary, as Spencer Davis died earlier this week. Today is a play-through of some of his music and the music of his contemporaries, exploring the sound of blue eyed soul and psychedelic 60s rock.