2020-10-23
Friday, October 23rd, 2020 - the Carol Kaye show! Carol Kaye is one of the most prolific recorded bass guitarists in rock and pop music, playing on an estimated 10,000 recordings in a career spanning over 50 years. She features heavily in the 2008 documentary The Wrecking Crew, and she is very likely the bass player that you never knew you knew. Today's show is a play-through of some of the contributions that she has made as a musician to popular music throughout the decades.