2020-11-01
Sunday, November 1st, 2020 - the funky as a motherfunker show! 174 shows in, and it seems like we're coming full circle as 🇬🇧 is about to enter lockdown. Who knows what's going to happen here (🇺🇸) next week.... So, shall we get the funk out of here for a little while? For two hours? Fancy getting funky? Funk it, why not?! Let's dance!