2020-11-12
Thursday, November 12th, 2020 - the nineties and noughties hip hop and rap show! Apologies for using the term "noughties" but the alliteration was too strong to resist.  I haven't got to play much of this genre on the show yet, but I LOVE it, even though I'm probably only slightly more knowledgeable on this than Jazz... Join me in an exploration, and grab a gin & juice!