2020-11-13
Friday, November 13th, 2020 - the mostly new music from mostly established artists show! Being Friday and all, it felt like the right time to get back to a little rock! More of a vibe-show today, than a strict theme, but as it shakes out, most of the music is new from 2020, from artists and bands that have been around for some time! Keep an ear out for the Morrissey track, it's excellent!