2020-11-14
Saturday, November 14th, 2020 - the more mellow multi-decade folk-rock show! My Dad shared a Fairport Convention track with me earlier in the week, and I asked him to follow it up with some songs of a similar vibe. Today's show is that playlist! Music from the 60s through to the 00s that you can sink back into and escape with.