2021-01-17
Sunday, January 17th, 2021 - the Steve Lillywhite show! Finally getting a chance to have a play-through of some of Steve Lillywhite's work. Punchy, punky, rocky, funky, slick, sparkling and sweet! Steve Lillywhite's career began in the early 1970s where he was a tape operator for Polygram. He was soon producing and contributing to some of the most well known and successful music of the following decades and continues to produce today having been awarded six Grammy awards and a CBE!