2021-01-24
Sunday, January 24th, 2021 - the United States musical travelogue show! Today's show was an alphabetical, autobiographical travelogue through the states I've lived in and visited. The rules were the songs had to have the State name in them, and for places I've lived I went a little deeper with some more songs with more geographical and biographical detail. From Dolly to Dylan, Ben Folds to Bill Monroe, Bloodhound Gang to Bill Withers, Elvis to Ellington, the music is as eclectic as all the places I've been lucky enough to see!