2021-01-31
Sunday, January 31st, 2021 - the eclectic, international effort show! Starting with a playlist curated almost to completion by Dave and Stanley Tongue, that they starting working on after listening to a show a few weeks ago. I selected a few extra tracks to bring us up to time for this play-through across a number of genres, with songs primarily from the 90s and 00s. There's also a special treat for the more discerning listener of the show with a premiere of a track just re-discovered on an old MiniDisc from the Baron Tongue himself!