2021-02-02
Tuesday, February 2nd, 2020 - the new music show! What's new in 2021? Most of the tracks today are, with some revisits to albums by artists who released new music towards the end of 2020. A variety of genres are represented today, as are a variety of countries. Keep an ear out for the Seaford Mods track, and if the sound from a mouth "fuck" upsets you, well, perhaps you have the wrong radio station, but it is sung in repetition by a 70 year old man during the last track. Who are we to judge?! 🤘🏻