2021-02-05
Friday, February 5th, 2020 - the perfect rocky Friday night getting ready show! I LOVE The Struts' new album Strange Days, and in particular the track Another Hit of Showmanship, and I'm always looking for a way to carry on that rocky, punchy, fun and funky vibe throughout a whole show. Since doing the new music shows on Tuesdays I've discovered so much new music and really caught up with what I've been missing out on over the past few years. So starting with The Struts, I began to build the playlist and to my great joy and amusement all the tracks I was picking were released in the last year, and many of the artists I've played on new music shows over the last 9 months or so. I pictured myself leaving work, heading to the local bar, loading the juke box up and sinking several strong beers to shake off the week. Cheers!