2021-02-11
Thursday, February 11th, 2021 - the supergroups show Pt. 2! Following on from yesterday's show, today we start in 1979 and listen through to the supergroups of the 1980s. A little louder, a little brighter, a little tighter, and a LOT more 80s. I think I was a little naive to think I could finish this journey in three shows. It's going to take a few more, and I'll spread them out across the next few weeks. More to come!