2021-02-13
Saturday, February 13th, 2021 - the hard-rock, blues-rock, groove-rock, funk-rock show! Slammin' Gladys released their new album Two yesterday and I listened to it so much I didn't do a show. I was looking for a way to represent the music in a way that didn't just lazily tie it back to the late wave of 80s rock/glam/hair bands that were washed away by the tsunami of grunge. Other than some of the tongue in cheek lyrics and overall playfulness of the first album I never considered them glam, so it was a shame when they and others lost label support and airplay for what is actually really good hard rock, groove rock etc. And it wasn't just them. Today's show is a play-through of that kind of blues-groove guitar driven music that hits hard and gets you air drumming, strutting, strumming, and singing along to the swagger.