2021-03-31
Wednesday, March 31st, 2021 - the most overlooked folk-rock LPs of the 1960s show! First thing I read this morning was a Rolling Stone article on Richard Thompson's new memoir Beeswing: Losing My Way and Finding My Voice. It tells a sad story of the fatal 1969 crash that killed the band's drummer Martin Lamble, and Thompson's girlfriend at the time Jeannie Franklin. The song I'll Keep It With Mine was played at Lamble's funeral. From here, I wanted to get into some of the deeper cuts of the folk-rock genre and found a superb article written by Richie Unterberger for Record Collector magazine in 2005. I played through some of the artists and built a playlist from here, and the show quickly evolved into a read-through of this article to accompany the selection of songs I had made. All credit to Richie Unterberger for this one!