2021-04-02
Friday, April 2nd, 2021 - the 80s soul, rock 'n roll show! Hugely inspired by the Tina Turner documentary "Tina!" on HBO Max (Sky Documentaries in the UK), I wanted to try to make up for underrating and under representing Tina Turner on the show so far. The documentary was a tough watch, but contained so much phenomenal footage of someone I now consider to be one of the greatest voices, and performers of all time. The 80s were a tricky time for a "sound" - so much stuff was overproduced and the sense of people recording together was lost a little as digital multi-tracking reduced the need to get things right in fewer takes. With that said though, the 80s sound is also really great, so I wanted to play-through some contemporaries of the era and just enjoy the music.