2021-04-10
Saturday, April 10th, 2021 - the perfect hangover day tunes show! The plan for this show was to be able to sit back and swaddle ourselves in some nostalgic, comforting 90s alternative, but it seems the universe had other ideas, and either Tidal, djay Pro AI, or the rattler decided they didn't want to play the songs I selected, so this show was an improvised exploration of what the world really wanted to hear!