2021-04-17
Saturday, April 17th, 2021 - the unashamedly 80s show! More technical gremlins – courtesy of MixCloud this time – delayed today's show until the unusual hour of 3:45PM, so instead of being an early 80s-shake-off-the-cobwebs type of show, it was a bit more of a late afternoon (delight?) type of show. Some singalong epic ballads, some upbeat dancers and some rock as well. Nice!