2021-04-22
Thursday, April 22nd, 2021 - the early 70s art/progressive/psychedelic/folk rock show! 1973 and 1974 are the sweet spots this evening. Starting off with some Barclay James Harvest, I always find myself wanting to extend and explore that ultra-mellow, super-smooth long form gentle progressive sound of theirs. Tonight was a meander through some albums I was mostly unfamiliar with from artists that I'm not. Not too much of any of the aforementioned genres, but a sprinkle of each to season your ears.