2021-04-24
Saturday, April 24th, 2021 - the yacht (canoe) rock show! Inspired by Dave and Craig's canoeing excursion on the Medway, I thought it would be nice to have some yacht rock to incongruously accompany them. As it turns out, it was a beautiful day everywhere and I ended up getting to enjoy a run along the Thames with Katherine, a BBQ in Hastings with the family, some ballet in the kitchen in Cheddar with Caroling, and the general feeling that all is right and sunny in the world! Thanks everyone :)