2021-04-27
Tuesday, April 27th, 2021 - the new music show! More new music, mostly singles from upcoming albums and EPs this time. I was very excited to wake up to find a new release from Counting Crows, and equally as excited to get to bookend the show with a Steve Cropper new release. In between are a variety of genres from artists new and old to tempt your ears with fresh material!