2021-05-12
Wednesday, May 12th, 2021 - the new, new music show! Normally scheduled for a Tuesday, today's new music show is a little different, but mainly behind the scenes in constructing the playlist and promoting it. Instead of playlisting the show immediately after I wake up, I listened a little more casually to music throughout the day and picked out tracks for the first hour that jumped out at me. The second half was put together more traditionally, but none of this matters to you! Most importantly I wanted to go back to a more easier going show, less focused on the promotion and more on the music and the experience. I had a lot of fun and made some new friends too! Hope you enjoy it! Squirrel flower.