2021-05-14
Friday, May 14th, 2021 - the road trip rock show! Another show for munching miles on the motorways, starting off with the effortlessly funky Lynyrd Skynyrd before settling in for the ride. This show isn't only for the US highways, it also features plenty of tunes from across the Atlantic. Time to get yourself a convertible, drop the top, select the gear and punch the gas (accelerator pedal... 😩). Watch out for the smokeys (police officers... 😖).