2021-05-16
Sunday, May 16th, 2021 - the ska, rocksteady, reggae, 2 tone, ska-punk show! I knew where I wanted to end up today, so I worked in reverse! I was really in the mood for some 2 tone and ska-punk, so I backed way up to play through some of the origins of the genre in Jamaica with ska, rocksteady and reggae before embarking on the trip to Coventry and then to the East and West coasts of the US. Starting off pretty laid back, the show builds throughout and you'll be pogo-ing and skanking right until the end as long as you've appropriately stretched first. I certainly discovered exactly how bouncy my chair is 😎