2021-05-26
Wednesday, May 26th, 2021 - the 1971 show! Inspired by watching the first few episodes of the new docuseries "1971: The Year That Music Changed Everything" today's show is a play-through of some of the featured artists and their peers. I've spent a few shows listening through and discussing the music of the 60s after watching the Woodstock and Laurel Canyon documentaries, and this is the logical and chronological next step in my musical wander through the decades.