2021-05-28
Friday, May 28th, 2021 - the last fifteen years or so of indie/alt rock show! Somewhere around 2004 there was a subtle shift in indie music and everything seemingly stepped up a gear and started falling more into place (with my taste). Oasis were becoming caricatures of themselves, and instead of bands trying to sound like them, there was a new post-punk influence creeping into the sound, and music started to become a little freer, a little more urgent and aggressive, and not to mention a little more tongue in cheek. I like it!