2021-05-29
Saturday, May 29th, 2021 - the "The" bands show! About 30 years ago there was a tiny article in a music magazine, or in the music section of one of the British newspapers, that talked about the trends in band names from really early on in the 1950s until (at that time) present day 1990s. I didn't keep the clipping, but I thought it was very astute, particularly how these trends were recycled and revisited by bands throughout the decades. Today's show is all bands that have "The" at the beginning. A dumb idea, but a really fun show!