2021-06-11
Friday, June 11th, 2021 - the early 80s pop rock show! Not an 80s metal show or an "80s rock" show, but rather the poppier, rockier early 80s big production, sometimes-soundtrack, radio friendly tunes from a selection of the bigger artists and albums of the time. As (mostly) always, not the obvious tracks, but a journey into some of the hidden gems that were sometimes overshadowed by the bigger hits.