2021-06-12
Saturday, June 12th, 2021 - the delicate dive into disco show! The Fratelli's have just released a cover of Yes Sir, I Can Boogie, and after listening to that this morning, that's the direction the show went. Slightly deeper cuts, and not all the usual obvious stuff, with a leaning towards the Euro beats of the time, today's show is a summery dancer perfect to accompany your day drinking dance parties!