2021-06-16
Wednesday, June 16th, 2021 - the 90s and 00s teen movie soundtrack vibe show! A little bit of post-grunge, a lotta bit of pop punk, a dash of alternative, a measure of rap rock, a sprinkle of hard rock and a drizzle of comedy rock. Today's show is all about the sound of the soundtrack to the teen movies of the era. Not strictly all teen movie tracks, and not a deep dive into teen movies, more a vibe inspired by the sound. Upbeat, amusing, easy to sing along too, and always fun!