2021-06-17
Thursday, June 17th, 2021 - the (mostly) hits from 1976 show! I like doing the deeper dives into specific years, it's almost like time travelling to the middle of America in a rental car, and riding the radio dial on a road-trip to nowhere. Today's show spans a number of genres, but doesn't drop too deep into any one of them. What did the world sound like in 1976? A fair amount of it likely sounded like this!