2021-06-18
Friday, June 18th, 2021 - the Lilith Fair 90s female vocalist show! It's been a few months since a show has focused purely on the female, so today's show is just that! Taking great inspiration from the performers that played the first Lilith Fair in 1997, and with a few contemporaries added in from my own personal favourites, today is a play-through of some hopefully not-so-obvious, and not-too-overplayed tracks from some of the biggest artists of the era.