2021-06-19
Saturday, June 19th, 2021 - the Philly soul sounds show! I've played a LOT of STAX, and a LOT of Motown on the show, but only really dipped my toe into the Philly soul sound, so today was an opportunity to educate myself. Starting off with a deep dive into Jimmy Bishop's Arctic records, and then moving into the Gamble and Huff Philadelphia International era, the tunes are upbeat, funky, luscious and loud!