2021-07-11
Sunday, July 11th, 2021 - the 1966 show! Let's not pretend that we're not all aware of a rather significant sporting event for English football (⚽️) fans today, but instead of playing the same old songs that every British radio station is likely playing, let's instead take a time machine back to 1966 to get a feeling of what the world of music sounded like when they won the World Cup in 1966. Deeper cuts and dustier selections, with some well-knowns throughout. Not a football show, but not an anti-football show either!