2021-07-18
Sunday, July 18th, 2021 - the 90s and 00s British groove rock and pop show! I'm always in the mood for a track with a little sassy groove, and today started out with Reef and Terrorvision bringing the British beer garden party vibes. We wandered around some deeper cuts of the decade before stepping into the early 00s and finishing off with some post-punk revival tracks and singalongs.