2021-08-12
Thursday, August 12th, 2021 - the new music show! Originally playlisted at the end of June this one has been in the hopper for some time, and when I listened through to some of the tracks again this morning I was completely in the mood for it! Soft and subtle, aggressive and energetic, sharp and stabby, grooving and bluesy, it moves through various genres and works well in a heavily humid 95 degrees (it may also work well at colder temperatures too!)