2021-08-13
Friday, August 13th, 2021 - the late 70s early 80s show! What was it like to be in your late teens, early 20s around 1978 - 1982, particularly if you enjoyed the vibe of the post-punk, power pop music that was coming up from everywhere?! It seems like it was a pretty good few years for music as long as the BBC didn't keep banning all your favourites! No censorship here. If the word fuck offends you, does the word duck? It's really similar!