2021-08-14
Saturday, August 14th, 2021 - the pop punk, power pop, pop rock, alternative show! I just bought tickets to see Bowling for Soup and it got me all excited for the early 00s rock scene. I was in my early mid 20s, just moved to Memphis and was surrounded by brilliant, interesting, creative people in an amazing city. It felt like an endless summer where every drive was with the top down, and everyone was working together on the same thing. This was the soundtrack!