2021-08-18
Wednesday, August 18th, 2021 - the UK top ten singles throughout 1985 show! A chronological deeper dive into some of the music that made 1985 sound like 1985. I was 5, and was just starting to LOVE music. I remember many of these from that time, and have purposefully skipped some of the more obvious or overplayed ones, in favour of the sounds that reside in the darkest depths of your subconscious ear. "If my calculations are correct, when this baby hits eighty-eight miles per hour... you're going to *hear some serious shit!"