2021-08-21
Saturday, August 21st, 2021 - the electronic, trip-hop, house, trance, UK garage, big beat show! I've been waiting a LONG time to do this show. My brother Nathan playlisted this for me a few months back and I've been holding off anxiously thinking I'd have to learn how to not ruin everything by doing a terrible job beat matching all the tracks. Instead of going that route, I just treated it like any other show and it was A LOT of fun. If clubbing was/is any part of your past or present this is a great throwback to the turn of the century!