2021-08-22
Sunday, August 22nd, 2021 - the 80s alty, indie, jangle, power-poppy show! Today's show is probably best described as a throwback to 1980s college radio in the US. Smooth, clean, crisp production, with alt agendas, upbeat, uplifting melodies, even if the message isn't always. Found at the "left side of the dial" tune in, turn it up, and let's be transported to a simpler time!