2021-08-24
Tuesday, August 24th, 2021 - the new music show! All new music released in the last month or so, this time including new releases of previously unheard material and well as a "covers" section. Sad day for music with the death of Charlie Watts, but as fate and the universe would have it, I had already playlisted a new/old track of theirs before I heard the news. Thanks for the tunes Charlie!