2021-08-26
Thursday, August 26th, 2021 - the sludgy, Seattle-y, grungey, alt rock show! If yesterday was glammy and glitzy, today's show is the antithesis, and although it wasn't 70s glam that grunge cleared the dance floor of, it was its similarly sparkly, upbeat, bubblegum, 80s offspring. Starting off with the "big five" today is a play-through of deeper album tracks from a variety of bands you will likely have heard of, and hopefully some that you haven't! Grab your flannel and your baggy jeans this is going to get sweaty.