2021-09-02
Thursday, September 2nd, 2021 - the Nashville tuning show! Nashville tuning, or high strung tuning is a great "trick" to add texture and character to a song often by adding it as a doubled track alongside a standard tuned guitar. Sweet, shimmering, and sparkly, it's a cool technique that cuts through the mix and adds a little extra sugar! After being inspired by a video my Dad sent me, I've ordered a couple of sets of strings to setup one of my guitars, and I wanted to explore how this method has been used across a number of genres in popular music beginning today in the late 1960s.