2021-09-05
Sunday, September 5th, 2021 - the 1955 show! Fire up the flux capacitor and let's head back in time to 1955 - a year I know most about from the film Back to The Future. It's also the year Steve Alemi was born, and today is his birthday so the show is a play-through of music written or released in that year. The rumblings of rock and roll rearing its head are present, as are some more traditional pop and jazz influenced songs. Listening through you can really get a sense of how much music is about to change... forever! Happy Birthday Steve! Spread out!