2021-09-08
Wednesday, September 8th, 2021 - the Tuesday new music show (on a Wednesday)! In a temporary adjustment to the usual schedule of new music being on a Tuesday, today's show is yesterday's playlist that I didn't get to broadcast. A little poppy, a little rocky, a little punky, a little pop punky, a little prog rocky, and just a sprinkle of djent. Join us for the regular play-through of new music released in the last few weeks.