2021-09-10
Friday, September 10th, 2021 - the early 90s hits and guilty pleasures show! There are so many songs and artists on today's show that I've not played yet that it almost feels cathartic to finally give them some air. I was ten and eleven when most of these came out, and a massive Guns N' Roses fan, but there were plenty of catchy pop tunes around that I liked and just didn't tell anyone. These are a those!