2021-09-12
Sunday, September 12th, 2021 - the birthday week through the years show! Number one singles from the UK or US through the years during the week of September 13th (my birthday) beginning in 1979 through to 2008. Sometimes people speak about the song that was number one on the day they were born, I wanted to have a listen to the music that was number one for each year! It's historical record so some of these are out of my hands. Some are absolute bangers, some are... well, listen and find out! :)