2021-09-21
Tuesday, September 21st, 2021 - the new music show! All new music released within the last 3 weeks or so, across the wide and wonderful genre of rock and indie. Not too heavy, not too gentle, and featuring a variety of veterans as well as more contemporary artists. Want to hear what's going on in new music now? Chuck this in your lugs and listen along with me!