2021-09-26
Sunday, September 26th, 2021 - the 1959 show! Selections from the Billboard Year End Hot 100 Singles of 1959. I wasn't planning on this show being overly US-centric, but when I started working though the list there was so much good stuff that I stayed on this side of the Atlantic. Lots of variety and plenty of hits that made it to the UK, as well as some fun facts about 1959 on both sides!