2021-09-28
Tuesday, September 28th, 2021 - the new music show! Rock, indie, blues, and folk in today's selection of new tracks all released this month. I'm starting to notice a trend towards slicker 80s production particularly in the drum sounds of some of these tracks. Listen out for The War on Drugs and Eddie Vedder as examples of this!