2021-09-30
Thursday, September 30th, 2021 - the Rolling Stone Best Songs of All Time show! No, it's not a show the length of Bible to include all 500 songs, it's just the top 27 from this year's list (and even that limited selection meant we ran over time!) All credit to Rolling Stone for the list and accompanying comments to today's play-through. What do you think of their curation?