2021-10-02
Saturday, October 2nd, 2021 - the Prince love-fest and more show! Following the terrific disappointment of losing power during yesterday's show, I was in the mood for some Prince to pull me out of the funk! Prince is a legend for all of his prolific bad-assery, and today's show is a spotlight on some of the perhaps lesser known tracks he's written, produced, and collaborated on during the 80s and early 90s. Following that catharsis is some power pop/rock/alternative classics to close it out!