2021-10-03
Sunday, October 3rd, 2021 - the hard rock power ballads from the 80s and 90s show! I've resisted playing many - if any - power ballads on the show until now, mainly because a lot of them are a little obvious and fairly overplayed. Other more obscure tracks don't tend to fit out of context, so there was only one thing for it: an unashamedly 80s and 90s power ballad show! This show is the sound of my headphones throughout the 90s and the posters on my wall. Warm your voice up and make sure to stretch before rocking out with air drums and guitar! Who wants a mixtape?