2021-10-07
Thursday, October 7th, 2021 - the 1982 UK and US Number Ones show! Today's show is a chronological play-through of number one singles in the UK and US for 1982. I was 2 turning 3 so I don't remember all of them, but it must have been just as I was starting to acknowledge music, because many of them ring with a clarity that can only be the result of a first-write to the brain matter! Tap into the time machine, and take yourself back almost four decades - there's some bangers on here!