2021-10-19
Tuesday, October 19th, 2021 - the new wave of classic rock show! Today's show is the Tuesday new music show with a little more flexibility in the definition of new. There are many new tracks - some released as recently as a week ago, others up to 6 years ago, but the resuscitation of the genre of melodic rock and AOR is more of the focus today. A favourite of mine through my teenage years, I was surprised recently to discover a band called Station, and then surprised again that they released a new album last week. I was even more surprised to discover there was a new melodic rock, or new wave of classic rock scene that I'd somehow missed beginning sometime around 2014-16. So today's show is an exploration into that!