2021-10-29
Friday, October 29th, 2021 - the album shopping in 1966 show! What would it have been like to be visiting music shops in 1966 every week of the year to buy albums, and what songs would I have chosen as favourites? What would have been yours? And what was the record shop you went to when you were buying music as a younger human?