2021-11-06
Saturday, November 6th, 2021 - the top tens of 1999 show! Selections from the US and UK top tens today, leaning heavily towards to the UK. There was a strong showing at the start of the year with lots of great tunes to choose from in February and March, before it thinned out into boy/girl band nonsense after the summer. Being 19 turning 20 in 1999, these were prime adventure years so the songs are seared into my memory and act as brilliant time travel devices to revisit the places and people I listened to them with.