2021-11-09
Tuesday, November 9th, 2021 - the new music in the folk/Americana genre show! Delayed a week, this is the playlist from last Tuesday that was postponed due to its inspiring effect on my own musical endeavours. With the clocks changing this weekend, listening to this show in the pitch black of 5pm just felt right. Nice and mellow, it allows you to drift off into the fiddles, banjos, and dobros, and imagine yourself under the big sky.