2021-11-10
Wednesday, November 10th, 2021 - the indie pop/alternative 00s show! Inspired by a re-introduction to the Ben Folds album Way To Normal, today's show is all music in a similar vein. Mostly mid to late 2000s when the production was slick, the songwriting was summery, and the world felt just a little bit simpler!