2021-11-12
Friday, November 12th, 2021 - the British post-punk revival show! Based on the first song, you may think you're listening to the wrong show, but to me it just fits, and then we're off to the races with a selection of not too obvious tunes from the mid to late 00s. The music that I missed out on when I moved to Memphis, but the music that I got to enjoy when I visited home. I've learned a lot about the original post-punk scene doing this show, and that makes this music all the more enjoyable. Crunchy, spiky, punky, sarcastic at times, this is the perfect throw back to having a good time in my 20s.