2021-11-21
Sunday, November 21st, 2021 - the Sunday Fun(k)day show! Today's show was a toss-up between a pretty edgy Hip-hop playlist or some super funky 70s sounds, and for the vibe today, the 70s won out! I love the groove of the Staple Singers I'll Take You There, but generally speaking I'm a bigger fan of 60s soul rather than the more extravagantly orchestrated 70s evolution. But, I LOVE some 70s funk, so I went with the groove of the Staple Singers and instead of building a soul show around it, I brought the funk! Not as rocky or aggressive as some of the previous funk shows, a little slower, and little more gentle, but be prepared to move!