2021-11-26
Friday, November 26th, 2021 - the mostly new hip hop and rap show! This is the playlist that I've been sitting on for the last week or so. It didn't feel right to do it during the week, and it's certainly not a weekend early show playlist as only one song lacks the E for explicit symbol, and only then because it's an instrumental. I do massively enjoy this genre of music, but I was most familiar with it from the 90s and early 00s. With the release of Snoop Dogg's new track Big Subwoofer, I was given the opportunity to have a dig around and I wasn't disappointed. Best to only have this one cranked when the kids aren't around!