2021-11-27
Saturday, November 27th, 2021 - the 80s, 90s, and 00s movie soundtracks show! I'm a massive movie fan, and equally keen on movie soundtracks, and have always enjoyed the shows I've done that are either dedicated to them, or feature specific songs. Today's show broadens the decades of past shows (mainly 80s) to include the 90s and 00s, which presents an enjoyable journey not only through a lot of my favourite movies growing up, but also an opportunity to hear the development of successful soundtrack songs.