2021-12-03
Friday, December 3rd, 2021 - the Friday folk show! How about some laid back indie pop/folk rock to soothe you into your Saturday and Sunday for folk's sake? All music released in the last 10 years or so - super mellow, rich, and beautifully produced - if you've had a bit of a week, this is an alternative to the usual get-up-and-party vibe of the normal Friday night shows. So just sit back and chill and let it all wash over you.