2021-12-04
Saturday, December 4th, 2021 - the hits of 2006 show! It was the year before the iPhone, the year that Twitter was launched, the year of the first Daniel Craig James Bond film, and the year that Pluto was planetarily demoted! There was also a lot of good music on the radio and today's show is a play-through of some of the more well known songs, along with facts from 2006 so you can be well prepared for your next pub quiz! All part of the service.