2021-12-18
Saturday, December 18th, 2021 - the week early Christmas show! Look, it's been a couple of years right? Life's hard, and normal feels like a distant memory even though we're still grasping for it at every opportunity. 18 months ago I did a Christmas in July show in part because it was funny, but also because Christmas is a seasonal comfort blanket, and even though sometimes it can take a minute to suppress your inner Grinch, it is (I hope) a time of year that brings with it fond memories as far back as you can remember. It's taken a while to feel Christmassy this year, but with the aid of the perfect hangover, and a large Covid coffee, we got there pretty quickly. Merry Christmas and Happy Holidays to everyone, whatever that means to you and however you choose to celebrate! Resist the eye roll, suppress the pessimism, let today's show be a time machine to your best Christmases, and a toast to those in the future!